FN-Scrapers-postgres_insert/tools/slack_scan/requirements.txt
astroid==1.6.2
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.3
PyYAML==3.12
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
slackclient==1.1.3
urllib3==1.22
websocket-client==0.47.0
wrapt==1.10.11
